feat: machine settings panel on dashboard * support for motion setting input * replaces previous machine limits panel on machine page
