+++ src/pages/Machine.vue
@@ -6,7 +6,6 @@
             </v-col>
             <v-col class="col-12 col-md-6 pt-0 pt-md-3">
                 <klippy-state-panel></klippy-state-panel>
-                <limits-panel></limits-panel>
                 <system-panel></system-panel>
                 <update-panel></update-panel>
                 <v-row>
@@ -26,14 +25,20 @@
 import {Component, Mixins} from 'vue-property-decorator'
 import BaseMixin from '@/components/mixins/base'
 import KlippyStatePanel from '@/components/panels/KlippyStatePanel.vue'
-import LimitsPanel from '@/components/panels/Machine/LimitsPanel.vue'
 import UpdatePanel from '@/components/panels/Machine/UpdatePanel.vue'
 import LogfilesPanel from '@/components/panels/Machine/LogfilesPanel.vue'
 import EndstopPanel from '@/components/panels/Machine/EndstopPanel.vue'
 import ConfigFilesPanel from '@/components/panels/Machine/ConfigFilesPanel.vue'
 import SystemPanel from '@/components/panels/Machine/SystemPanel.vue'
 @Component({
-    components: {SystemPanel, ConfigFilesPanel, EndstopPanel, LogfilesPanel, UpdatePanel, LimitsPanel, KlippyStatePanel}
+    components: {
+        SystemPanel,
+        ConfigFilesPanel,
+        EndstopPanel,
+        LogfilesPanel,
+        UpdatePanel,
+        KlippyStatePanel
+    }
 })
 export default class PageMachine extends Mixins(BaseMixin) {
